Use C aggregate initialization to avoid warnings
If we try to initialize each element explicitly in a struct during brace
initialization but we make a mistake (or when the struct changes) the
compiler will complain with "missing braces around initializer" (GCC).
However this is unnecessary: simply using {0} will force aggregate
initialization of the entire structure no matter how complex (this is
required by C99).